requirements.txt
beautifulsoup4==4.10.0
Django==3.2.11
lxml==4.7.1
python-decouple==3.5
requests==2.27.1
rstr==3.0.0
